spotless apply

+++ app/src/main/java/com/example/androiddevchallenge/NavGraph.kt
@@ -58,7 +58,7 @@ fun NavGraph(startDestination: String = MainDestinations.WELCOME) {
 
 class MainActions(navController: NavController) {
 
-    val actionStartToLogin: (Int) -> Unit = {
+    val actionStartToLogin: () -> Unit = {
         navController.navigate(MainDestinations.LOGIN)
     }
